More comments & cleanup
