+++ src/main/scala/workflow/graph/Identity.scala
@@ -1,0 +1,15 @@
+package workflow.graph
+
+import org.apache.spark.rdd.RDD
+
+import scala.reflect.ClassTag
+
+/**
+ * This transformer performs a no-op on its input.
+ *
+ * @tparam T Type of the input and, by definition, output.
+ */
+case class Identity[T : ClassTag]() extends Transformer[T,T] {
+  override protected def singleTransform(in: T): T = in
+  override protected def batchTransform(in: RDD[T]): RDD[T] = in
+}
