+++ src/main/scala/workflow/graph/Pipeline.scala
@@ -297,117 +297,6 @@ object Transformer {
   }
 }
 
-
-abstract class Estimator[A, B] extends EstimatorOperator {
-  final def fit(data: RDD[A]): Pipeline[A, B] = {
-    fit(PipelineRDDUtils.rddToPipelineDatasetOut(data))
-  }
-  final def fit(data: PipelineDatasetOut[A]): Pipeline[A, B] = {
-    // Remove the data sink,
-    // Then insert this estimator into the graph with the data as the input
-    val curSink = data.getGraph.getSinkDependency(data.getSink)
-    val (estGraph, estId) = data.getGraph.removeSink(data.getSink).addNode(this, Seq(curSink))
-
-    // Now that the estimator is attached to the data, we need to build a pipeline DAG
-    // that applies the fit output of the estimator. We do this by creating a new Source in the DAG,
-    // Adding a delegating transformer that depends on the source and the label estimator,
-    // And finally adding a sink that connects to the delegating transformer.
-    val (estGraphWithNewSource, sourceId) = estGraph.addSource()
-    val (almostFinalGraph, delegatingId) = estGraphWithNewSource.addNode(new DelegatingOperator, Seq(estId, sourceId))
-    val (newGraph, sinkId) = almostFinalGraph.addSink(delegatingId)
-
-    // Because pipeline construction is incremental, we make sure to add the states of the data,
-    // removing the graph ids that no longer exist.
-    val newState = data.getState - data.getSink
-    new ConcretePipeline(new GraphExecutor(newGraph, newState), sourceId, sinkId)
-  }
-
-  final override private[graph] def fitRDDs(inputs: Seq[DatasetExpression]): TransformerOperator = {
-    fitRDD(inputs.head.get.asInstanceOf[RDD[A]])
-  }
-  protected def fitRDD(data: RDD[A]): Transformer[A, B]
-}
-
-abstract class LabelEstimator[A, B, L] extends EstimatorOperator {
-  final def fit(data: RDD[A], labels: PipelineDatasetOut[L]): Pipeline[A, B] = {
-    fit(PipelineRDDUtils.rddToPipelineDatasetOut(data), labels)
-  }
-  final def fit(data: PipelineDatasetOut[A], labels: RDD[L]): Pipeline[A, B] = {
-    fit(data, PipelineRDDUtils.rddToPipelineDatasetOut(labels))
-  }
-  final def fit(data: RDD[A], labels: RDD[L]): Pipeline[A, B] = {
-    fit(PipelineRDDUtils.rddToPipelineDatasetOut(data), PipelineRDDUtils.rddToPipelineDatasetOut(labels))
-  }
-
-  final def fit(data: PipelineDatasetOut[A], labels: PipelineDatasetOut[L]): Pipeline[A, B] = {
-    // Add the data input and the labels inputs into the same Graph
-    val (dataAndLabels, labelSourceMapping, labelNodeMapping, labelSinkMapping) =
-      data.getGraph.addGraph(labels.getGraph)
-
-    // Remove the data sink & the labels sink,
-    // Then insert this label estimator into the graph with the data & labels as the inputs
-    val dataSink = dataAndLabels.getSinkDependency(data.getSink)
-    val labelsSink = dataAndLabels.getSinkDependency(labelSinkMapping(labels.getSink))
-    val (estimatorWithInputs, estId) = dataAndLabels
-      .removeSink(data.getSink)
-      .removeSink(labelSinkMapping(labels.getSink))
-      .addNode(this, Seq(dataSink, labelsSink))
-
-    // Now that the labeled estimator is attached to the data & labels, we need to build a pipeline DAG
-    // that applies the fit output of the estimator. We do this by creating a new Source in the DAG,
-    // Adding a delegating transformer that depends on the source and the label estimator,
-    // And finally adding a sink that connects to the delegating transformer.
-    val (estGraphWithNewSource, sourceId) = estimatorWithInputs.addSource()
-    val (almostFinalGraph, delegatingId) = estGraphWithNewSource.addNode(new DelegatingOperator, Seq(estId, sourceId))
-    val (newGraph, sinkId) = almostFinalGraph.addSink(delegatingId)
-
-    // Because pipeline construction is incremental, we make sure to add the states of the data & labels,
-    // updating graph ids for the new graph, and removing all graph ids that no longer exist.
-    val graphIdMappings: Map[GraphId, GraphId] = labelSourceMapping ++ labelNodeMapping ++ labelSinkMapping
-    val newState = data.getState ++
-      labels.getState.map(x => (graphIdMappings(x._1), x._2)) -
-      data.getSink -
-      labelSinkMapping(labels.getSink)
-
-    // Finally, we construct a new pipeline w/ the new graph & new state.
-    new ConcretePipeline(new GraphExecutor(newGraph, newState), sourceId, sinkId)
-  }
-
-  final override private[graph] def fitRDDs(inputs: Seq[DatasetExpression]): TransformerOperator = {
-    fitRDDs(inputs(0).get.asInstanceOf[RDD[A]], inputs(1).get.asInstanceOf[RDD[L]])
-  }
-  protected def fitRDDs(data: RDD[A], labels: RDD[L]): Transformer[A, B]
-
-}
-
-/**
- * This transformer performs is a no-op on its input.
- *
- * @tparam T Type of the input and, by definition, output.
- */
-case class Identity[T : ClassTag]() extends Transformer[T,T] {
-  override protected def singleTransform(in: T): T = in
-  override protected def batchTransform(in: RDD[T]): RDD[T] = in
-}
-
-/**
- * Caches an RDD at a given point within a Pipeline. Follows Spark's lazy evaluation conventions.
- *
- * @param name An optional name to set on the cached output. Useful for debugging.
- * @tparam T Type of the input to cache.
- */
-case class Cacher[T: ClassTag](name: Option[String] = None) extends Transformer[T,T] with Logging {
-  override protected def batchTransform(in: RDD[T]): RDD[T] = {
-    logInfo(s"CACHING ${name.getOrElse(in.id)}")
-    name match {
-      case Some(x) => in.cache().setName(x)
-      case None => in.cache()
-    }
-  }
-
-  override protected def singleTransform(in: T): T = in
-}
-
 case class Checkpointer[T : ClassTag](path: String) extends Estimator[T, T] {
   override protected def fitRDD(data: RDD[T]): Transformer[T, T] = {
     data.saveAsObjectFile(path)
